geronimo/requirements.txt
Django==1.7.7
ipaddress==1.0.7
